Shippable version of Messier Handbook.
The only drawbacks that I notice are that some of the Messier objects
are overexposed in DSS (M78, M83), and some of them are misaligned due
to RA / Dec problems in KStars (M41, M21 etc).

Also, some DSS images have plate defects (M106, M37? etc) and it would
be best to use different wavelength surveys for those objects.